% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aug_admm_genlasso.R
\name{aug_admm_genlasso}
\alias{aug_admm_genlasso}
\title{Augmented ADMM Solver for Generalized Lasso Problems}
\usage{
aug_admm_genlasso(
  A,
  b,
  D = diag(length(b)),
  M,
  lambda = 1,
  rho = 1,
  abstol = 1e-04,
  reltol = 0.01,
  maxiter = 1000
)
}
\arguments{
\item{A}{Data matrix.}

\item{b}{Response vector.}

\item{D}{Regularization matrix, defaults to the identity matrix of the length of \code{b}.}

\item{M}{Augmentation matrix used in the augmented ADMM.}

\item{lambda}{Regularization parameter.}

\item{rho}{Augmentation parameter for ADMM.}

\item{abstol}{Absolute tolerance level for convergence.}

\item{reltol}{Relative tolerance level for convergence.}

\item{maxiter}{Maximum number of iterations for the algorithm.}
}
\value{
A list containing the following components:
\itemize{
\item \code{x}: The solution vector.
\item \code{x_iter}: Solution vector at each iteration.
\item \code{history}: A data frame containing the objective value, primal and dual norms,
and epsilon values for primal and dual convergence checks at each iteration.
}
}
\description{
Solves the generalized lasso problem using the augmented Alternating Direction Method of Multipliers (ADMM).
This function extends the standard ADMM approach by incorporating an additional matrix 'M', which can lead to more efficient computations in certain scenarios.
}
\examples{
# Generate synthetic data
n <- 200; m <- 100; p <- 0.1
A <- matrix(rnorm(m * n), nrow = m)
b <- matrix(rnorm(m))
D <- diag(n)
M <- diag(n) # Example augmentation matrix

# Solve using aug_admm_genlasso
result <- aug_admm_genlasso(A, b, D, M)

}
